+++ crates/server/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "server"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+repository.workspace = true
+
+[dependencies]
+storage = { path = "../storage" }
