+++ crates/storage/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "storage"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+repository.workspace = true
+
+[dependencies]
